requirement.txt
ansi==0.1.3
asn1crypto==0.24.0
bcrypt==3.1.4
beautifulsoup4==4.6.3
bottle==0.12.13
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
colorlog==3.1.4
cryptography==2.4.2
Cython==0.29
daemonize==2.4.7
dnspython==1.15.0
dnspython3==1.15.0
errbot==5.2.0
future==0.17.1
get==2018.11.19
idna==2.7
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.0
paramiko==2.4.2
pkg-resources==0.0.0
post==2018.11.20
public==2018.11.20
pyasn1==0.4.4
pycparser==2.19
Pygments==2.2.0
pygments-markdown-lexer==0.1.0.dev39
PyNaCl==1.3.0
pyOpenSSL==18.0.0
python-telegram-bot==11.1.0
query-string==1.0.2
request==2018.11.20
requests==2.20.1
rocket-errbot==1.2.5
six==1.11.0
urllib3==1.24.1
waitress==1.1.0
WebOb==1.8.4
WebTest==2.0.32
Yapsy==1.12.0
